Open-BFME5: the index-buffer lock classes hold a DX8 device mutex
~WriteLockClass at 0x00138840, 113 bytes exact, and with it the identity of
four callees this file's remaining bodies all depend on.

BFME guards DX8 buffer access with a real device mutex where Zero Hour had a
thread assert that compiles to nothing. 0x0011F520 takes it -- a 20-second
WaitForSingleObject on the handle at 0x00DEC598, reporting through
Debug::SkipNext if it times out -- and 0x00120F50 releases it, decrementing the
recursion count at 0x00DEDA8C under the critical section at 0x00DEC540.
WriteLockClass acquires in its constructor and releases in its destructor, so
the lock is held for the life of the lock object. All four addresses are read
from retail REL32 displacements and are self-verifying now that a body around
them matches.

The shape matters as much as the calls. Retail's destructor carries a C++ EH
frame with the state running 0 -> -1 and `this` spilled to [esp+4] for the
funclet: that is an RAII guard, not a bare call pair. Two details finish it:

  * the guard carries an owner pointer initialised from `this`, which is the
    `mov [esp+4],esi` a memberless guard does not emit; and
  * its destructor is INLINE, so the release is reached by a direct `call` with
    no `lea ecx` -- declared-not-defined instead costs four bytes and misses.

The unit was also calling its local forceinline BFME_DX8_ErrorCode here. Retail
calls Log_DX8_ErrorCode out of line at 0x0011F450, which is where the DXER tag
and the debug stream actually live; the inline copy belongs only where retail
really inlines it.

Also in this commit, the second half of the index_count widening:
Get_Index_Count returns `unsigned` in BFME too. The WriteLockClass constructor
reads the count with a plain `mov eax,[eax+0Ch]` and doubles it, where an
`unsigned short` accessor forces a 16-bit load and a movzx. That constructor is
now down to its EH frame alone -- 109 of 171 bytes with every instruction in
the right order -- and the frame is a member-construction state, which says
WriteLockClass holds the guard as a member rather than as a local.

+++ reference/shims/indexbuffercount/dx8indexbuffer.h
@@ -68,7 +68,10 @@ public:
 	void Copy(unsigned int* indices,unsigned start_index,unsigned index_count);
 	void Copy(unsigned short* indices,unsigned start_index,unsigned index_count);
 
-	inline unsigned short Get_Index_Count() const { return index_count; }
+	// Also 32-bit in BFME: the WriteLockClass constructor at 0x00138790 reads
+	// the count with a plain `mov eax,[eax+0Ch]` and doubles it, where an
+	// unsigned short accessor forces a 16-bit load and a movzx.
+	inline unsigned Get_Index_Count() const { return index_count; }
 
 	inline unsigned Type() const { return type; }
